Handle partially locked request in CL req r/w unlock
Allowing request cacheline lock to be called multiple
times on a single request. This is going to be usefull
for upcomming eviction improvements, where request will
first have evicted cachelines assigned to it in a locked
state and then the hit cachelines will subsequently be
locked to complete cachelines preparation.

Signed-off-by: Adam Rutkowski <adam.j.rutkowski@intel.com>